requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy==2.0.23
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pandas==2.1.2
matplotlib==3.8.1
